refactor: add login requirements to ReservationController

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,8 +1,11 @@
 package nextstep.reservation;
 
+import nextstep.member.Member;
+import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
 import nextstep.support.DuplicateEntityException;
+import nextstep.support.NotExistEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
@@ -14,14 +17,16 @@ public class ReservationService {
     public final ReservationDao reservationDao;
     public final ThemeDao themeDao;
     public final ScheduleDao scheduleDao;
+    public final MemberDao memberDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao, MemberDao memberDao) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.scheduleDao = scheduleDao;
+        this.memberDao = memberDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, String username) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -32,9 +37,14 @@ public class ReservationService {
             throw new DuplicateEntityException();
         }
 
+        Member member = memberDao.findByUsername(username);
+        if (member == null) {
+            throw new NullPointerException();
+        }
+
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                member.getName()
         );
 
         return reservationDao.save(newReservation);
@@ -49,10 +59,19 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(String username, Long id) {
+        Member member = memberDao.findByUsername(username);
+        if (member == null) {
+            throw new NullPointerException();
+        }
+
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
+        }
+
+        if (!reservation.getName().equals(member.getName())) {
+            throw new NotExistEntityException();
         }
 
         reservationDao.deleteById(id);
